fix mixin, typings for mixin, and test robot can move around map
