+++ client/tests/robot.ts
@@ -1,7 +1,7 @@
 // @ts-nocheck
 import mocha from "mocha";
 import { expect } from "chai";
-import Robot from "../src/robot/logic";
+import Robot from "../src/robot";
 import PhysicalWorld from '../src/world';
 import { WORLD as TestWorld } from './fixtures';
